fix: output as paperless.html
- fix `generateId` so we can set a `slug` property in md frontmatter to pick a specific filename for output
- set `slug: paperless.html` in the report to ensure it is generated as `paperless.html`
- delete redirect as it's now built to the right place.

The redirect method worked, but it ment that the user always sees the redirect as github was preferring to load paperless.html first for the path `/paperless` with no tailing slash.

With this fix we can now pick an output filename if we want to, or let it be the post filename.

I'm trying to recreate this specific bit of jekyll automagic here, but should decide if we want to put add at least a "year of publish" path prefix on our posts so the come out like `/2025/hett-show`, so that if (when) we go to HETT again, we don't have to aribtrarily name the file something other to avoid clashing when the date stamp is removed from the `yyyy-mm-dd-filename.md`.

License: MIT
Signed-off-by: Oli Evans <oli@zilla.org.uk>